requirements.txt
streamlit
openai
numpy
openai
python-dotenv
pyaudio
opencv-python
panda
pydantic

librosa
matplotlib
pydub
plotly
mediapipe